+++ src/components/DomainsCombinationsChecker/DomainsCombinationsChecker.module.css
@@ -1,4 +1,4 @@
-.AdvancedDomainsChecker {
+.DomainsCombinationsChecker {
     /*/
     outline: 1px dotted rgb(123, 255, 0);
     /**/
