+++ src/components/DomainsCombinationsChecker/DomainsCombinationsChecker.tsx
@@ -1,15 +1,15 @@
 import { useState } from 'react';
 import spaceTrim from 'spacetrim';
 import { string_domain, string_tdl } from '../../utils/typeAliases';
-import styles from './AdvancedDomainsChecker.module.css';
-import { DomainsStatusList } from './DomainsStatusList/DomainsStatusList';
-import { createAllPermutationsOf } from './utils/createAllPermutationsOf';
-import { createAllSubsetsOf } from './utils/createAllSubsetsOf';
+import styles from './DomainsCombinationsChecker.module.css';
+import { DomainsStatusList } from '../Domains/DomainsStatusList/DomainsStatusList';
+import { createAllPermutationsOf } from '../Domains/utils/createAllPermutationsOf';
+import { createAllSubsetsOf } from '../Domains/utils/createAllSubsetsOf';
 
 /**
  * Renders a domain checker with advanced options and patterns
  */
-export function AdvancedDomainsChecker() {
+export function DomainsCombinationsChecker() {
     const [names, setNames] = useState<Array<string_domain>>(['web', 'gpt']);
     const [tdls, setTdls] = useState<Array<string_tdl>>(['com', /*'org', 'io', 'net',*/ 'cz']);
 
@@ -26,7 +26,7 @@ export function AdvancedDomainsChecker() {
     const sortedDomains = uniqueDomains.sort((a, b) => a.length - b.length);
 
     return (
-        <div className={styles.AdvancedDomainsChecker}>
+        <div className={styles.DomainsCombinationsChecker}>
             <div className={styles.patterns}>
                 <textarea
                     className={styles.pattern}
