--- src/pages/experiments/loading.tsx
@@ -1,5 +1,0 @@
-import { TasksInProgress } from '../../components/TaskInProgress/TasksInProgress';
-
-export default function TestLoadingPage() {
-    return <TasksInProgress />;
-}
