+++ src/pages/experiments/loading/loading.tsx
@@ -1,0 +1,5 @@
+import { TasksInProgress } from '../../../components/TaskInProgress/TasksInProgress';
+
+export default function TestLoadingPage() {
+    return <TasksInProgress />;
+}
